pyproject.toml
[tool.poetry]
name = "similarium"
version = "0.1.0"
description = "A Slack bot to play Similarium together as a team"
authors = ["Axel <dev@absalon.is>"]
license = "GNU GPLv3"

[tool.poetry.dependencies]
python = "~3.10"
slack-sdk = "^3.15.2"
slack-bolt = "^1.13.0"
aiohttp = "^3.8.1"
SQLAlchemy = "^1.4.35"
rich = "^12.2.0"
gensim = {version = "^4.1.2", optional = true}
numpy = {version = "^1.22.3", optional = true}
scipy = {version = "^1.8.0", optional = true}
toml = "^0.10.2"
aiosqlite = {version = "^0.17.0", optional = true}
asyncpg = {version = "^0.25.0", optional = true}
python-dateutil = "^2.8.2"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"
pyright = "^1.1.237"
flake8 = "^4.0.1"
flake8-builtins = "^1.5.3"
flake8-isort = "^4.1.1"
flake8-bugbear = "^22.3.23"
flake8-black = "^0.3.2"
ipdb = "^0.13.9"
pdbpp = "^0.10.3"
pytest = "^7.1.1"
pytest-clarity = "^1.0.1"
rich = "^12.2.0"
ipython = "^8.2.0"
sqlalchemy-stubs = "^0.4"
pytest-asyncio = "^0.18.3"

[tool.poetry.extras]
dump = ["numpy", "gensim", "scipy"]
sqlite = ["aiosqlite"]
postgres = ["asyncpg"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
extra_standard_library = ["typing_extensions"]

[tool.black]
target-version = ["py310"]

[tool.pyright]
pythonVersion = "3.10"
include = ["src", "tests"]
